+++ components/mip-shell-xiaoshuo/common/constant-config.js
