+++ components/mip-shell-xiaoshuo/example/mipx-xiaoshuo-1-3.html
@@ -156,6 +156,9 @@
       padding: 10px;
       border: 1px solid #999;
     }
+    .mip-shell-xiaoshuo-container {
+      opacity: 0;
+    }
   </style>
   <noscript>
     <style mip-officialrelease="">
@@ -188,7 +191,7 @@
                     "chapter": 1,
                     "page": 3,
                     "isLastPage": false,
-                    "chapterName": "第一章第一节"
+                    "chapterName": "第一章第三节"
                 },
                 "previousPage": {
                     "url": "",
@@ -210,7 +213,7 @@
   <div class="click-cursor mip-shell-xiaoshuo-container" on="click:xiaoshuo-shell.showShellFooter">
     <span class="link click-cursor" on="click:xiaoshuo-shell.showShellCatalog">打开目录</span>
     <span class="link click-cursor" on="click:xiaoshuo-shell.showShellFooter">打开设置</span>
-    <a class="link" href="./mipx-xiaoshuo-2-1.html?crid=2&pg=1" mip-link="">下一页</a>
+    <a  cache-first class="link" href="./mipx-xiaoshuo-2-1.html?crid=2&pg=1" mip-link="">下一页</a>
     <h3>第1章 灵魂重生(3)</h3>
     <p>
       “贱人，你竟敢背叛我！”
@@ -406,7 +409,7 @@
     </p>
   </div>
 </div>
-<a class="link" href="./mipx-xiaoshuo-1-2.html" mip-link="">下一页</a>
+<a  cache-first class="link" href="./mipx-xiaoshuo-1-2.html" mip-link="">下一页</a>
 <footer id="footer">
   小说网 &copy; 2018 - 冀ICP备xxxx号
 </footer>
@@ -418,7 +421,7 @@
     }
   </script>
 </mip-custom>
-<mip-shell-xiaoshuo mip-shell="" id="xiaoshuo-shell">
+<mip-shell-xiaoshuo prerender mip-shell="" id="xiaoshuo-shell">
   <script type="application/json">
     {
       "routes": [
