+++ components/mip-shell-xiaoshuo/example/mipx-xiaoshuo-3-1.html
@@ -163,6 +163,9 @@
       padding: 10px;
       border: 1px solid #999;
     }
+    .mip-shell-xiaoshuo-container {
+      opacity: 0;
+    }
 
   </style>
   <noscript>
@@ -197,7 +200,7 @@
           "chapter": 3,
           "page": 1,
           "isLastPage": false,
-          "chapterName": "第一章第一节"
+          "chapterName": "第三章第一节"
         },
         "previousPage": {
           "url": "",
@@ -219,7 +222,7 @@
 <body>
 <span class="link click-cursor" on="click:xiaoshuo-shell.showShellCatalog">打开目录</span>
 <span class="link click-cursor" on="click:xiaoshuo-shell.showShellFooter">打开设置</span>
-<a class="link" href="./mipx-xiaoshuo-2-1.html?crid=2&pg=1" mip-link="">上一页</a>
+<a  cache-first class="link" href="./mipx-xiaoshuo-2-1.html?crid=2&pg=1" mip-link="">上一页</a>
 <div class="click-cursor" on="click:xiaoshuo-shell.showShellFooter">
   <h3>第3章 灵魂重生(1)</h3>
   <p>
@@ -415,7 +418,7 @@
     陆宇眼神坚定，他是魂天师，就算是废武魂，他也一样能搅动天地，称雄万世！
   </p>
 </div>
-<a class="link" href="./mipx-xiaoshuo-1-2.html" mip-link="">下一页</a>
+<a  cache-first class="link" href="./mipx-xiaoshuo-1-2.html" mip-link="">下一页</a>
 <footer id="footer">
   小说网 &copy; 2018 - 冀ICP备xxxx号
 </footer>
